requirements.txt
openai
flask
azure-search-documents